# Use concurrent set for global symbol table