--- a/concurrent_set.c
+++ b/concurrent_set.c
@@ -119,7 +119,7 @@ concurrent_set_try_resize_without_locking(VALUE old_set_obj, VALUE *set_obj_ptr)
         RUBY_ASSERT(key != CONCURRENT_SET_MOVED);
 
         if (key < CONCURRENT_SET_SPECIAL_VALUE_COUNT) continue;
-        if (rb_objspace_garbage_object_p(key)) continue;
+        if (!RB_SPECIAL_CONST_P(key) && rb_objspace_garbage_object_p(key)) continue;
 
         VALUE hash = RUBY_ATOMIC_VALUE_LOAD(entry->hash);
         if (hash == 0) {
@@ -231,7 +231,7 @@ rb_concurrent_set_find_or_insert(VALUE *set_obj_ptr, VALUE key, void *data)
             VALUE curr_hash = RUBY_ATOMIC_VALUE_LOAD(entry->hash);
             if ((curr_hash == hash || curr_hash == 0) && set->funcs->cmp(key, curr_key)) {
                 // We've found a match.
-                if (UNLIKELY(rb_objspace_garbage_object_p(curr_key))) {
+                if (UNLIKELY(!RB_SPECIAL_CONST_P(curr_key) && rb_objspace_garbage_object_p(curr_key))) {
                     // This is a weakref set, so after marking but before sweeping is complete we may find a matching garbage object.
                     // Skip it and mark it as deleted.
                     RUBY_ATOMIC_VALUE_CAS(entry->key, curr_key, CONCURRENT_SET_DELETED);