This PR switches the global symbol table and allows for symbol lookup and creation of dynamic symbols without needing the global VM lock. Creating static symbols still require the global VM lock but I will be investigating how to make that lock-free in the future.

Here's a microbenchmark demonstrating the speedup:

```ruby
ARGV[0].to_i.times.map do
  Ractor.new do
    1_000_000.times do |i|
      "hello#{i}".to_sym
    end
  end
end.map(&:value)
```

Results:

<img width="884" height="547" alt="chart" src="https://github.com/user-attachments/assets/75f84817-5876-4a0d-bd7e-cee1d5e94c11" />

Raw results:

| Ractor count | Branch (s) | Master (s) |
|--------------|------------|------------|
| 1            | 0.364      | 0.401      |
| 2            | 0.555      | 1.149      |
| 3            | 0.583      | 3.890      |
| 4            | 0.680      | 3.288      |
| 5            | 0.789      | 5.107      |